[Refactor] 서비스 로직 계층분리
