+++ darwin/src/app/[lng]/components/Product/useProductCard.ts
@@ -3,7 +3,7 @@ import { toast } from "react-toastify";
 import { Product } from "@/domain/product/Product";
 import { useModalState } from "@/util/modal/useModalState";
 
-import { SuccessToast } from "@/app/components/ui/Toast";
+import { SuccessToast } from "@/components/ui/Toast";
 import { useI18n } from "@/app/i18n/I18nProvider";
 
 export const useProductCard = ({
