+++ darwin/src/app/[lng]/layout.tsx
@@ -7,7 +7,7 @@ import dynamic from "next/dynamic";
 import { Global } from "@emotion/react";
 import { dir } from "i18next";
 
-import { toastStyles } from "../components/ui/Toast/toastStyles";
+import { toastStyles } from "@/components/ui/Toast/toastStyles";
 import I18nProvider from "../i18n/I18nProvider";
 import { ChakraLayoutProvider, QueryProviders } from "../providers";
 import Header from "./components/Header/Header";
