+++ darwin/src/components/ui/FilePreview/index.tsx
